build: update test config
