+++ package.json
@@ -28,13 +28,13 @@
     "site:preview": "cd site && vite build --mode preview",
     "postsite:preview": "cp _site/index.html _site/404.html",
     "test": "npm run test:unit && npm run test:snap",
-    "test:unit": "vitest run --config site/vite.config.js",
-    "test:update": "vitest run --config site/vite.config.js -u && npm run test:snap-update",
-    "test:unit-dev": "vitest --config site/vite.config.js",
-    "test:unit-gui": "vitest --config site/vite.config.js --ui",
-    "test:unit-coverage": "vitest run --config site/vite.config.js --coverage",
-    "test:snap": "cross-env NODE_ENV=test-snap vitest run --config site/vite.config.js",
-    "test:snap-update": "cross-env NODE_ENV=test-snap vitest run --config site/vite.config.js -u",
+    "test:unit": "vitest run",
+    "test:update": "vitest run -u && npm run test:snap-update",
+    "test:unit-dev": "vitest",
+    "test:unit-gui": "vitest --ui",
+    "test:unit-coverage": "vitest run --coverage",
+    "test:snap": "cross-env NODE_ENV=test-snap vitest run",
+    "test:snap-update": "cross-env NODE_ENV=test-snap vitest run -u",
     "test:e2e": "cypress run --config-file scripts/test/cypress.json",
     "test:e2e-gui": "cypress open --config-file scripts/test/cypress.json",
     "test:demo": "node scripts/test/generate-demo-test.js",
