+++ scripts/test/snap-setup.js
