+++ site/vite.config.js
@@ -12,36 +12,6 @@ const publicPathMap = {
   production: 'https://static.tdesign.tencent.com/mobile-vue/',
 };
 
-// 单元测试相关配置
-const testConfig = {
-  include:
-    /**
-     * 快照涵盖 `__test__/*.test.jsx`
-     * 生成demo测试文件： npm run test:demo
-     * 生成快照：npm run test:snap || npm run test:snap-update
-     *
-     * 测试用例检测 && GUI && 覆盖率报告，仅涵盖 `__test__/index.test.jsx`
-     * 单测： npm run test:unit
-     * 覆盖率报告： npm run test:unit-coverage
-     * GUI：npm run test:unit-gui
-     */
-    process.env.NODE_ENV === 'test-snap'
-      ? ['src/**/__test__/*.{test,spec}.{js,mjs,cjs,ts,mts,cts,jsx,tsx}']
-      : ['src/**/__test__/index.{test,spec}.{js,mjs,cjs,ts,mts,cts,jsx,tsx}'],
-  globals: true,
-  environment: 'jsdom',
-  testTimeout: 5000,
-  setupFiles: process.env.NODE_ENV === 'test-snap' ? path.resolve(__dirname, '../scripts/test/test-setup.js') : '',
-  transformMode: {
-    web: [/\.[jt]sx$/],
-  },
-  coverage: {
-    reporter: ['text', 'json', 'html'],
-    reportsDirectory: 'test/unit/coverage',
-    exclude: ['src/**.{js,ts}', 'src/_common/**', 'src/shared/**', 'src/**/{__test__,demos,style}/**'],
-  },
-};
-
 const isCustomElement = (tag) => tag.startsWith('td-');
 
 export default ({ mode }) => {
@@ -84,6 +54,5 @@ export default ({ mode }) => {
       }),
       createTDesignPlugin(),
     ],
-    test: testConfig,
   });
 };
